requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.3.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.11.0
django-import-export==2.8.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
et-xmlfile==1.1.0
ez-setup==0.9
factory-boy==3.2.1
Faker==13.3.4
filelock==3.6.0
identify==2.4.12
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
MarkupPy==1.14
MarkupSafe==2.1.1
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.22.3
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.9
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.18.1
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
boto3
djangorestframework-simplejwt
django_heroku
django-silk
sentry-sdk
gunicorn
cloudinary
python-docx
dj-stripe
stripe
django-ckeditor
htmldocx
django-rest-passwordreset
django-sendgrid-v5
django-hashid-field
simple_salesforce
python-decouple
phonenumbers
django-storages
boto3
django-cloudinary-storage
django-anymail
selenium
django-celery-results
django-celery-beat
redis
Pillow